Create README.md
I have a problem like this, I was going to convert my game to the old version in Roblox Studio and I tried it, but then when I tried to convert it to the latest version from the website, it didn't work. I even turned off and on team creation and tried it again, it didn't work.